requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altgraph==0.17.5
anyio==4.12.1
argcomplete==3.6.3
attrs==26.1.0
audioop-lts==0.2.2
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
blivet==3.12.1
borgbackup==1.4.4
Brlapi==0.8.7
Brotli==1.2.0
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-360.1-build/cockpit-360.1/tmp/wheel/cockpit-360.1-py3-none-any.whl#sha256=7ce868b58854dd32e336f23a9779654e5b0bfa20956c83c8331f25e6d706112c
crypt_r==3.13.1
cryptography==46.0.7
cupshelpers==1.0
customtkinter==5.2.2
Cython==3.1.3
darkdetect==0.8.0
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
discord.py==2.7.1
distro==1.9.0
dnf==4.24.0
evdev==1.9.3
fedora-third-party==0.10
file-magic==0.4.0
Flask==3.1.3
frozenlist==1.8.0
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
invoke==3.0.3
iso639==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.6
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.18.0
llfuse==1.5.2
louis==3.33.0
lutris==0.5.22
lxml==6.0.1
MarkupSafe==3.0.3
moddb==0.12.0
msgpack==1.1.0
multidict==6.7.1
mutagen==1.47.0
nftables==0.1
numexpr==2.12.1
numpy==2.3.5
olefile==0.47
packaging==25.0
paramiko==4.0.0
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
playsound==1.2.2
ply==3.11
plyer==2.1.0
productmd==1.50
propcache==0.4.1
protobuf==3.19.6
protontricks==1.13.1
psutil==7.0.0
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycryptodomex==3.23.0
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.54.5
pyinotify==0.9.6
pyinstaller==6.20.0
pyinstaller-hooks-contrib==2026.4
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
PyNaCl==1.6.2
pyparted==3.13.0
pypresence==4.3.0
pyproj==3.7.2
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2026.2.28
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rpm==6.0.1
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==82.0.1
simpleline==1.9.0
six==1.17.0
sos==4.11.1
soupsieve==2.8
systemd-python==235
typing_extensions==4.15.0
urllib3==2.6.3
vdf==3.4
websockets==15.0.1
Werkzeug==3.1.8
xkbregistry==0.3
yarl==1.23.0
yt-dlp==2026.3.17
zstandard==0.25.0
